// src/app/interfaces/productResponse.interface.ts
// Generated by https://quicktype.io

export interface ProductResponse {
  metadata:        Metadatum[];
  productResponse: ProductResponseClass;
}

export interface Metadatum {
  date: string;
  code: string;
  type: string;
}

export interface ProductResponseClass {
  products: Product[];
}

export interface Product {
  id:       number;
  name:     string;
  price:    number;
  account:  number;
  category: Category;
  picture:  string;
}

export interface Category {
  id:          number;
  name:        string;
  description: string;
}